set logical order of methods.  still need to consolidate code and implement the interface
