+++ Project02Eclipse/results.out
@@ -1,23 +1,23 @@
-anxiety=5
-arrhythmias=3
-blindness=1
-blurred vision=5
-constricted pupils=3
-cough=6
-dialated pupils=4
-dizziness=5
-dry mouth=8
-fever=7
-headache=3
-high blood pressure=10
-inflamation=7
-insomnia=4
-low blood pressure=4
-nausea=5
-rapid heart rate=1
-rash=4
-shortness of breath=3
-stiff neck=4
-stomach pain=3
-tremor=4
-water retention=1
+anxiety=10
+arrhythmias=6
+blindness=2
+blurred vision=10
+constricted pupils=6
+cough=12
+dialated pupils=8
+dizziness=10
+dry mouth=16
+fever=14
+headache=6
+high blood pressure=20
+inflamation=14
+insomnia=8
+low blood pressure=8
+nausea=10
+rapid heart rate=2
+rash=8
+shortness of breath=6
+stiff neck=8
+stomach pain=6
+tremor=8
+water retention=2
